Fix undefined behavior in 'pty' and '.sh.match' (#708)
This commit fixes a few errors that occur when ksh is built with
-fsanitize=address,undefined.

src/lib/libast/include/proc.h:
- Use long typecasts in PROC_op* macros to make the bit shifts safe.
  This fixes a shift overflow in the expansion of PROC_FD_CTTY():
      src/cmd/builtin/pty.c:356:18: runtime error: left shift of 12
      by 28 places cannot be represented in type 'int'

src/cmd/ksh93/sh/init.c:
- Ensure mp->nodes is not NULL before making use of it. This change
  fixes two errors under ASan that occur during the exit and glob
  regression tests:
      src/cmd/ksh93/sh/init.c:868:6: runtime error: applying
      non-zero offset 18446744073709551600 to null pointer
      src/cmd/ksh93/sh/init.c:831:5: runtime error: applying
      non-zero offset 18446744073709551600 to null pointer

+++ src/lib/libast/include/proc.h
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -64,8 +64,8 @@
 
 #define PROC_fd_ctty	0xc
 
-#define PROC_op1(o,a)	(((o)<<(2*PROC_ARG_BIT))|((a)&((PROC_ARG_NULL<<PROC_ARG_BIT)|PROC_ARG_NULL)))
-#define PROC_op2(o,a,b)	(((o)<<(2*PROC_ARG_BIT))|(((b)&PROC_ARG_NULL)<<PROC_ARG_BIT)|((a)&PROC_ARG_NULL))
+#define PROC_op1(o,a)	(((long)(o)<<(2*PROC_ARG_BIT))|((long)(a)&((PROC_ARG_NULL<<PROC_ARG_BIT)|PROC_ARG_NULL)))
+#define PROC_op2(o,a,b)	(((long)(o)<<(2*PROC_ARG_BIT))|(((long)(b)&PROC_ARG_NULL)<<PROC_ARG_BIT)|((long)(a)&PROC_ARG_NULL))
 
 #define PROC_FD_CLOSE(p,f)	PROC_op2(PROC_fd_dup|(f),p,PROC_ARG_NULL)
 #define PROC_FD_CTTY(f)		PROC_op1(PROC_fd_ctty,f)
